+++ opm/output/eclipse/EclipseReader.cpp
@@ -23,8 +23,8 @@
 #include <ert/util/ert_unique_ptr.hpp>
 
 #include <opm/output/eclipse/EclipseReader.hpp>
-#include <opm/output/Cells.hpp>
-#include <opm/output/Wells.hpp>
+#include <opm/output/data/Cells.hpp>
+#include <opm/output/data/Wells.hpp>
 #include <opm/parser/eclipse/EclipseState/EclipseState.hpp>
 #include <opm/parser/eclipse/EclipseState/InitConfig/InitConfig.hpp>
 #include <opm/parser/eclipse/EclipseState/IOConfig/IOConfig.hpp>
