+++ opm/output/eclipse/EclipseReader.hpp
@@ -3,8 +3,8 @@
 
 #include <utility>
 
-#include <opm/output/Cells.hpp>
-#include <opm/output/Wells.hpp>
+#include <opm/output/data/Cells.hpp>
+#include <opm/output/data/Wells.hpp>
 
 namespace Opm {
