+++ tests/test_Restart.cpp
@@ -27,8 +27,8 @@
 
 #include <opm/output/eclipse/EclipseWriter.hpp>
 #include <opm/output/eclipse/EclipseReader.hpp>
-#include <opm/output/Cells.hpp>
-#include <opm/output/Wells.hpp>
+#include <opm/output/data/Cells.hpp>
+#include <opm/output/data/Wells.hpp>
 #include <opm/parser/eclipse/Deck/Deck.hpp>
 #include <opm/parser/eclipse/EclipseState/Grid/EclipseGrid.hpp>
 #include <opm/parser/eclipse/EclipseState/IOConfig/IOConfig.hpp>
